when removing, only descend into children that can contain the node to remove
A node's bbox is the union of its children's bboxes.  When searching for
a node with a specific extent, the node can only be in children who's
bboxes strictly contain the node's extent.  We can therefore replace the
intersection test with a stricter contains test.  This improves the
speed of remove by about 50%.